break up js
